separated out sessions-related code
Note: #defining ENABLE_SESSIONS to be 1 will result in untested behavior.
