+++ moo.h
@@ -124,12 +124,6 @@
 #define DEBUG_PIN5_LOW
 #endif
 
-#if ENABLE_SESSIONS
-void initialize_sessions();
-void handle_session_timeout();
-int bitCompare(unsigned char *, unsigned short, unsigned char *,
-               unsigned short, unsigned short);
-#endif // ENABLE_SESSIONS
 void setup_to_receive();
 void sleep();
 unsigned short is_power_good();
