+++ mymoo.h
@@ -69,6 +69,9 @@
 #define ENABLE_SLOTS 			0
 #define ENABLE_SESSIONS			0
 #define ENABLE_HANDLE_CHECKING          0 // not implemented yet ...
+#if ENABLE_SESSIONS
+#include "sessions.h"
+#endif
 ////////////////////////////////////////////////////////////////////////////////
 
 ////////////////////////////////////////////////////////////////////////////////
